refactor(coin-solana): move HTTP calls to `network` folder
