+++ .changeset/selfish-walls-wait.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-solana": minor
+---
+
+refactor(coin-solana): move HTTP calls to `network` folder
