+++ libs/coin-modules/coin-solana/src/broadcast.ts
@@ -1,7 +1,7 @@
 import { log } from "@ledgerhq/logs";
 import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
 import type { Account, Operation, SignedOperation } from "@ledgerhq/types-live";
-import { ChainAPI } from "./api";
+import { ChainAPI } from "./network";
 import { SolanaTxConfirmationTimeout, SolanaTxSimulationFailedWhilePendingOp } from "./errors";
 import { BlockhashWithExpiryBlockHeight } from "@solana/web3.js";
 
@@ -23,7 +23,7 @@ export const broadcastWithAPI = async (
       rawData?.recentBlockhash as BlockhashWithExpiryBlockHeight,
     );
     return patchOperationWithHash(operation, txSignature);
-  } catch (e: any) {
+  } catch (e) {
     // heuristics to make some errors more user friendly
     if (e instanceof Error) {
       log("broadcast-error", e.message);
