+++ libs/coin-modules/coin-solana/src/network/chain/instruction/stake/types.ts
