+++ libs/coin-modules/coin-solana/src/network/chain/program/constants.ts
